ENTB-19 Improve UserServiceImpl
